pyproject.toml
[tool.poetry]
name = "command-agent"
version = "0.1.0"
description = "An agent capable of memorizing reusable procedure as \"command\", built on langchain"
authors = ["rynsuke <ey.nosukeru@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "command_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.0.168"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
notion-client = "^2.0.0"
pinecone-client = "^2.2.1"
tiktoken = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
